+++ client/packages/coldchain/src/Equipment/Components/index.ts
@@ -1,3 +1,2 @@
-export * from './AssetLogPanel';
 export * from './Status';
 export * from './FileList';
