+++ client/packages/coldchain/src/Equipment/DetailView/UpdateStatusButton.tsx
@@ -6,28 +6,17 @@ import {
   useDialog,
   DialogButton,
   useNotification,
-  TabContext,
   Box,
-  useTabs,
-  useDebounceCallback,
   DetailContainer,
   InsertAssetLogInput,
   FnUtils,
-  ClickableStepper,
   UserPermission,
   useAuthContext,
   useIsGapsStoreOnly,
 } from '@openmsupply-client/common';
-import { StatusTab } from './StatusTab';
-import { UploadTab } from './UploadTab';
+import { StatusForm, Draft } from './StatusForm';
 import { useAssets } from '../api';
-import { Draft } from '../Components';
 import { Environment } from '@openmsupply-client/config/src';
-
-enum Tabs {
-  Status = 'Status',
-  Upload = 'UploadFiles',
-}
 
 const getEmptyAssetLog = (assetId: string) => ({
   id: FnUtils.generateUUID(),
@@ -41,9 +30,7 @@ export const UpdateStatusButtonComponent = ({
 }) => {
   const onClose = () => {
     setDraft(getEmptyAssetLog(assetId ?? 'closed'));
-    onChangeTab(Tabs.Status);
   };
-  const { currentTab, onChangeTab } = useTabs(Tabs.Status);
   const t = useTranslation();
   const { Modal, hideDialog, showDialog } = useDialog({ onClose });
   const { error, success, info } = useNotification();
@@ -59,10 +46,6 @@ export const UpdateStatusButtonComponent = ({
       showDialog();
     } else info(t('error.no-asset-edit-permission'))();
   };
-
-  const onNext = useDebounceCallback(() => {
-    onChangeTab(Tabs.Upload);
-  }, []);
 
   const onOk = async () => {
     await insertLog(draft)
@@ -94,41 +77,9 @@ export const UpdateStatusButtonComponent = ({
       .catch(e => error(`${t('error.unable-to-save-log')}: ${e.message}`)());
   };
 
-  const logSteps = [
-    {
-      description: '',
-      label: t('label.status'),
-      tab: Tabs.Status,
-      clickable: true,
-    },
-    {
-      description: '',
-      label: t('label.upload-files'),
-      tab: Tabs.Upload,
-    },
-  ];
-
-  const getActiveStep = () => {
-    const step = logSteps.find(step => step.tab === currentTab);
-    return step ? logSteps.indexOf(step) : 0;
-  };
-
-  const isInvalid = () => !draft?.id || !draft?.assetId || !draft?.status;
-
   const onChange = (patch: Partial<InsertAssetLogInput>) => {
     if (!draft) return;
     setDraft({ ...draft, ...patch });
-  };
-
-  const onClickStep = (tabName: string) => {
-    switch (tabName) {
-      case Tabs.Upload:
-        onChangeTab(tabName as Tabs);
-        break;
-      case Tabs.Status:
-        onChangeTab(tabName as Tabs);
-        break;
-    }
   };
 
   useEffect(() => {
@@ -151,19 +102,9 @@ export const UpdateStatusButtonComponent = ({
           />
         }
         okButton={
-          currentTab === Tabs.Upload ? (
-            <DialogButton variant="ok" onClick={onOk} />
-          ) : undefined
+          <DialogButton variant="ok" onClick={onOk} />
         }
-        nextButton={
-          currentTab === Tabs.Status ? (
-            <DialogButton
-              variant="next-and-ok"
-              onClick={onNext}
-              disabled={isInvalid()}
-            />
-          ) : undefined
-        }
+
       >
         <DetailContainer paddingTop={1}>
           <Box
@@ -178,23 +119,10 @@ export const UpdateStatusButtonComponent = ({
               },
             }}
           >
-            <ClickableStepper
-              activeStep={getActiveStep()}
-              steps={logSteps}
-              onClickStep={onClickStep}
+            <StatusForm
+              draft={draft}
+              onChange={onChange}
             />
-            <TabContext value={currentTab}>
-              <StatusTab
-                draft={draft}
-                value={Tabs.Status}
-                onChange={onChange}
-              />
-              <UploadTab
-                draft={draft}
-                value={Tabs.Upload}
-                onChange={onChange}
-              />
-            </TabContext>
           </Box>
         </DetailContainer>
       </Modal>
